+++ Sources/GenerateMedia/GenerateMedia.swift
@@ -96,7 +96,11 @@ private struct ViewShot: View {
     let scheme: ColorScheme
 
     var body: some View {
-        IconPickerView(icon: self.$icon, color: self.$color)
+        IconPickerView(
+            icon: self.$icon,
+            color: self.$color,
+            colors: Array(IconPickerColor.all.prefix(6)),
+            allowsCustomColor: true)
             .preferredColorScheme(self.scheme)
     }
 }
@@ -108,7 +112,11 @@ private struct RowShot: View {
 
     var body: some View {
         Form {
-            IconPickerRow(icon: self.$icon, color: self.$color)
+            IconPickerRow(
+                icon: self.$icon,
+                color: self.$color,
+                colors: Array(IconPickerColor.all.prefix(6)),
+                allowsCustomColor: true)
         }
         .formStyle(.grouped)
         .preferredColorScheme(self.scheme)